# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "guest-place",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/content/entries/{id}": {
      "get": {
        "tags": [
          "Content"
        ],
        "operationId": "get_entry",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Entry ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ContentEntry"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Content"
        ],
        "operationId": "delete_entry",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Entry ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Value"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Content"
        ],
        "operationId": "update_entry",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Entry ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateContentEntryDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ContentEntry"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/v1/system/alert": {
      "post": {
        "tags": [
          "System"
        ],
        "operationId": "send_system_alert",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendAlertDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Alert broadcasted"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse_ContentEntry": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "id",
              "schema_id",
              "slug",
              "data",
              "status",
              "i18n",
              "created_at",
              "updated_at"
            ],
            "properties": {
              "created_at": {
                "type": "string",
                "format": "date-time"
              },
              "data": {
                "type": "object"
              },
              "i18n": {
                "type": "object"
              },
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "published_at": {
                "type": [
                  "string",
                  "null"
                ],
                "format": "date-time"
              },
              "published_by": {
                "type": [
                  "string",
                  "null"
                ],
                "format": "uuid"
              },
              "schema_id": {
                "type": "string",
                "format": "uuid"
              },
              "slug": {
                "type": "string"
              },
              "status": {
                "$ref": "#/components/schemas/ContentEntryStatus"
              },
              "updated_at": {
                "type": "string",
                "format": "date-time"
              }
            }
          },
          "errors": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "messages": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "ApiResponse_Value": {
        "type": "object",
        "properties": {
          "data": {},
          "errors": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "messages": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "ContentEntryStatus": {
        "type": "string",
        "enum": [
          "draft",
          "review",
          "published"
        ]
      },
...
      "SendAlertDTO": {
        "type": "object",
        "required": [
          "title",
          "message"
        ],
        "properties": {
          "level": {
            "type": [
              "string",
              "null"
            ]
          },
          "message": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        }
      },
...
      "UpdateContentEntryDTO": {
        "type": "object",
        "properties": {
          "data": {},
          "i18n": {},
          "published_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "published_by": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "slug": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ContentEntryStatus"
              }
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "gp_apikey"
      },
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "bearer_auth": []
    }
  ],
...
}